requirements.txt
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
